Move keyboard shortcut loading to C++
