+++ src/openrct2/config/Config.h
@@ -220,4 +220,10 @@ extern "C"
 {
     bool config_open(const utf8 * path);
     bool config_save(const utf8 * path);
+    void config_get_default_path(utf8 *outPath, size_t size);
+    void config_set_defaults();
+    void config_release();
+    bool config_open_default();
+    bool config_save_default();
+    bool config_find_or_browse_install_directory();
 }
