+++ src/openrct2/interface/keyboard_shortcut.h
@@ -29,4 +29,72 @@ void keyboard_shortcut_handle(sint32 key);
 void keyboard_shortcut_handle_command(sint32 shortcutIndex);
 void keyboard_shortcut_format_string(char *buffer, size_t size, uint16 shortcutKey);
 
+void config_reset_shortcut_keys();
+bool config_shortcut_keys_load();
+bool config_shortcut_keys_save();
+
+typedef struct shortcut_entry {
+	uint8 key;
+	uint8 modifier;
+} shortcut_entry;
+
+extern uint16 gShortcutKeys[SHORTCUT_COUNT];
+
+enum {
+	SHORTCUT_CLOSE_TOP_MOST_WINDOW,
+	SHORTCUT_CLOSE_ALL_FLOATING_WINDOWS,
+	SHORTCUT_CANCEL_CONSTRUCTION_MODE,
+	SHORTCUT_PAUSE_GAME,
+	SHORTCUT_ZOOM_VIEW_OUT,
+	SHORTCUT_ZOOM_VIEW_IN,
+	SHORTCUT_ROTATE_VIEW_CLOCKWISE,
+	SHORTCUT_ROTATE_VIEW_ANTICLOCKWISE,
+	SHORTCUT_ROTATE_CONSTRUCTION_OBJECT,
+	SHORTCUT_UNDERGROUND_VIEW_TOGGLE,
+	SHORTCUT_REMOVE_BASE_LAND_TOGGLE,
+	SHORTCUT_REMOVE_VERTICAL_LAND_TOGGLE,
+	SHORTCUT_SEE_THROUGH_RIDES_TOGGLE,
+	SHORTCUT_SEE_THROUGH_SCENERY_TOGGLE,
+	SHORTCUT_INVISIBLE_SUPPORTS_TOGGLE,
+	SHORTCUT_INVISIBLE_PEOPLE_TOGGLE,
+	SHORTCUT_HEIGHT_MARKS_ON_LAND_TOGGLE,
+	SHORTCUT_HEIGHT_MARKS_ON_RIDE_TRACKS_TOGGLE,
+	SHORTCUT_HEIGHT_MARKS_ON_PATHS_TOGGLE,
+	SHORTCUT_ADJUST_LAND,
+	SHORTCUT_ADJUST_WATER,
+	SHORTCUT_BUILD_SCENERY,
+	SHORTCUT_BUILD_PATHS,
+	SHORTCUT_BUILD_NEW_RIDE,
+	SHORTCUT_SHOW_FINANCIAL_INFORMATION,
+	SHORTCUT_SHOW_RESEARCH_INFORMATION,
+	SHORTCUT_SHOW_RIDES_LIST,
+	SHORTCUT_SHOW_PARK_INFORMATION,
+	SHORTCUT_SHOW_GUEST_LIST,
+	SHORTCUT_SHOW_STAFF_LIST,
+	SHORTCUT_SHOW_RECENT_MESSAGES,
+	SHORTCUT_SHOW_MAP,
+	SHORTCUT_SCREENSHOT,
+
+	// New
+	SHORTCUT_REDUCE_GAME_SPEED,
+	SHORTCUT_INCREASE_GAME_SPEED,
+	SHORTCUT_OPEN_CHEAT_WINDOW,
+	SHORTCUT_REMOVE_TOP_BOTTOM_TOOLBAR_TOGGLE,
+	SHORTCUT_SCROLL_MAP_UP,
+	SHORTCUT_SCROLL_MAP_LEFT,
+	SHORTCUT_SCROLL_MAP_DOWN,
+	SHORTCUT_SCROLL_MAP_RIGHT,
+	SHORTCUT_OPEN_CHAT_WINDOW,
+	SHORTCUT_QUICK_SAVE_GAME,
+	SHORTCUT_SHOW_OPTIONS,
+	SHORTCUT_MUTE_SOUND,
+	SHORTCUT_WINDOWED_MODE_TOGGLE,
+	SHORTCUT_SHOW_MULTIPLAYER,
+	SHORTCUT_PAINT_ORIGINAL_TOGGLE,
+	SHORTCUT_DEBUG_PAINT_TOGGLE,
+	SHORTCUT_SEE_THROUGH_PATHS_TOGGLE,
+
+	SHORTCUT_COUNT
+};
+
 #endif
