+++ src/openrct2/libopenrct2.vcxproj
@@ -75,6 +75,7 @@
     <ClCompile Include="config\Config.cpp" />
     <ClCompile Include="config\IniReader.cpp" />
     <ClCompile Include="config\IniWriter.cpp" />
+    <ClCompile Include="config\KeyboardShortcuts.cpp" />
     <ClCompile Include="FileClassifier.cpp" />
     <ClCompile Include="network\ServerList.cpp" />
     <ClCompile Include="ParkImporter.cpp" />
@@ -91,7 +92,6 @@
     <ClCompile Include="cmdline\ScreenshotCommands.cpp" />
     <ClCompile Include="cmdline\SpriteCommands.cpp" />
     <ClCompile Include="cmdline_sprite.c" />
-    <ClCompile Include="config.c" />
     <ClCompile Include="core\Console.cpp" />
     <ClCompile Include="core\Diagnostics.cpp" />
     <ClCompile Include="core\File.cpp" />
@@ -440,7 +440,6 @@
     <ClInclude Include="cmdline\CommandLine.hpp" />
     <ClInclude Include="cmdline_sprite.h" />
     <ClInclude Include="common.h" />
-    <ClInclude Include="config.h" />
     <ClInclude Include="core\Collections.hpp" />
     <ClInclude Include="core\Console.hpp" />
     <ClInclude Include="core\Diagnostics.hpp" />
